requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
groq
tiktoken
langgraph
langchain-text-splitters
ddgs
langchain-ollama
numpy
transformers
langgraph-checkpoint-sqlite
duckduckgo-search
faiss-cpu
streamlit
langchain-tavily
python-dotenv
pypdf
ollama
scikit-learn
langchain-chroma
youtube-transcript-api
chromadb
tavily
langchain-nvidia-ai-endpoints
yt-dlp